requirements.txt
langchain==0.3.25
langchain-openai==0.3.16
langchain-community==0.3.24
langchain-chroma==0.2.4
openai==1.82.0
chromadb>=1.0.9
python-dotenv==1.1.0
gradio==5.31.0
streamlit==1.56.0
